chore: reduce quality budget footprint
